// src/common/enums/building-type.enum.ts
export enum BuildingType {
  HOSPITAL = 'hospital',
  SCHOOL = 'school',
  UNIVERSITY = 'university',
  EDUCATIONAL = 'educational',
  PUBLIC = 'public',
  MOSQUE = 'mosque',
  CHURCH = 'church',
  RELIGIOUS = 'religious',
  POLICE = 'police',
  EMBASSY = 'embassy',
  PORT = 'port',
  ARMY = 'army',
}